refactor: enforce Bounded Context boundaries
Extract shared query infrastructure:
- Move MetaCompare and SortDirection from PostType\Repository
  to shared Query namespace (BackTo\Framework\Query)
- Both PostType and Taxonomy now import from the shared namespace

Decouple PostMeta from PostType bounded context:
- Create PostReferenceInterface in PostMeta\Contracts as ACL
- PostMeta no longer imports anything from PostType
- PostInterface extends PostReferenceInterface for compatibility

Dependency map after refactoring:
- PostMeta → (no external BC deps)
- Taxonomy → Query (shared), no PostType deps
- PostType → Query (shared), PostMeta (MetaKey VO only)
