+++ src/PostMeta/Entity/PostMeta.php
@@ -6,12 +6,12 @@ namespace BackTo\Framework\PostMeta\Entity;
 
 use BackTo\Framework\PostMeta\Contracts\PostMetaInterface;
 use BackTo\Framework\PostMeta\ValueObject\MetaKey;
-use BackTo\Framework\PostType\Contracts\PostInterface;
+use BackTo\Framework\PostMeta\Contracts\PostReferenceInterface;
 
 final class PostMeta implements PostMetaInterface
 {
     private int $id;
-    private ?PostInterface $post = null;
+    private ?PostReferenceInterface $post = null;
     private int $postId = 0;
     private MetaKey $metaKey;
     private mixed $metaValue;
@@ -27,7 +27,7 @@ final class PostMeta implements PostMetaInterface
         return $this;
     }
 
-    public function getPost(): PostInterface
+    public function getPost(): PostReferenceInterface
     {
         if ($this->post === null) {
             throw new \LogicException('Post has not been set on this PostMeta.');
@@ -35,7 +35,7 @@ final class PostMeta implements PostMetaInterface
         return $this->post;
     }
 
-    public function setPost(PostInterface $post): PostMetaInterface
+    public function setPost(PostReferenceInterface $post): PostMetaInterface
     {
         $this->post = $post;
         $this->postId = $post->getId() ?? 0;
