+++ src/PostType/Tests/PostSpecificationTest.php
@@ -7,7 +7,7 @@ namespace BackTo\Framework\PostType\Tests;
 use BackTo\Framework\PostMeta\ValueObject\MetaKey;
 use BackTo\Framework\PostType\Entity\PostStatus;
 use BackTo\Framework\PostType\Factory\PostFactory;
-use BackTo\Framework\PostType\Repository\MetaCompare;
+use BackTo\Framework\Query\MetaCompare;
 use BackTo\Framework\PostType\Repository\PostQueryBuilder;
 use BackTo\Framework\PostType\Specification\AndPostSpecification;
 use BackTo\Framework\PostType\Specification\PostsByAuthor;
