+++ src/Taxonomy/Repository/TermQueryBuilder.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace BackTo\Framework\Taxonomy\Repository;
 
-use BackTo\Framework\PostType\Repository\MetaCompare;
-use BackTo\Framework\PostType\Repository\SortDirection;
+use BackTo\Framework\Query\MetaCompare;
+use BackTo\Framework\Query\SortDirection;
 use BackTo\Framework\Taxonomy\Contracts\TermInterface;
 use BackTo\Framework\Taxonomy\Factory\TermFactory;
 use BackTo\Framework\Taxonomy\Specification\TermSpecification;
